refactor: Extract service protection types to their own module
